Changes in response to review comments
