+++ infrastructure/terraform/components/branch/module_amplify_branch.tf
@@ -20,6 +20,5 @@ module "amplify_branch" {
   environment_variables = {
     NOTIFY_SUBDOMAIN              = var.environment
     NEXT_PUBLIC_BASE_PATH         = "/templates~${local.normalised_branch_name}"
-    MAX_TTL_DURATION_IN_SECONDS = 432000 # 5 days in seconds
   }
 }
