+++ next.config.js
@@ -6,8 +6,6 @@ const nextConfig = {
   basePath,
   env: {
     basePath,
-    MAX_TTL_DURATION_IN_SECONDS:
-      process.env.MAX_TTL_DURATION_IN_SECONDS ?? '432000', // 5 days in seconds
   },
 
   experimental: {
