--- src/app/create-nhs-app-template/[templateId]/page.tsx
@@ -1,23 +1,0 @@
-'use server';
-
-import { CreateNhsAppTemplate } from '@forms/CreateNhsAppTemplate/CreateNhsAppTemplate';
-import { PageProps } from '@utils/types';
-import { getTemplate } from '@utils/form-actions';
-import { redirect, RedirectType } from 'next/navigation';
-import { validateNHSAppTemplate } from '@utils/validate-template';
-
-const CreateNhsAppTemplatePage = async ({
-  params: { templateId },
-}: PageProps) => {
-  const template = await getTemplate(templateId);
-
-  const validatedTemplate = validateNHSAppTemplate(template);
-
-  if (!validatedTemplate) {
-    redirect('/invalid-template', RedirectType.replace);
-  }
-
-  return <CreateNhsAppTemplate initialState={validatedTemplate} />;
-};
-
-export default CreateNhsAppTemplatePage;
