+++ src/app/create-text-message-template/page.tsx
@@ -1,4 +1,4 @@
-import { CreateSmsTemplate } from '@forms/CreateSmsTemplate/CreateSmsTemplate';
+import { SmsTemplateForm } from '@forms/SmsTemplateForm/SmsTemplateForm';
 import { SMSTemplate, Draft } from '@utils/types';
 import { TemplateType } from '@utils/enum';
 
@@ -10,7 +10,7 @@ const CreateSMSTemplatePage = async () => {
     message: '',
   };
 
-  return <CreateSmsTemplate initialState={initialState} />;
+  return <SmsTemplateForm initialState={initialState} />;
 };
 
 export default CreateSMSTemplatePage;
