+++ src/app/edit-nhs-app-template/[templateId]/page.tsx
@@ -1,0 +1,23 @@
+'use server';
+
+import { NhsAppTemplateForm } from '@forms/NhsAppTemplateForm/NhsAppTemplateForm';
+import { PageProps } from '@utils/types';
+import { getTemplate } from '@utils/form-actions';
+import { redirect, RedirectType } from 'next/navigation';
+import { validateNHSAppTemplate } from '@utils/validate-template';
+
+const CreateNhsAppTemplatePage = async ({
+  params: { templateId },
+}: PageProps) => {
+  const template = await getTemplate(templateId);
+
+  const validatedTemplate = validateNHSAppTemplate(template);
+
+  if (!validatedTemplate) {
+    redirect('/invalid-template', RedirectType.replace);
+  }
+
+  return <NhsAppTemplateForm initialState={validatedTemplate} />;
+};
+
+export default CreateNhsAppTemplatePage;
