--- src/components/forms/CreateNhsAppTemplate/CreateNhsAppTemplate.tsx
@@ -1,103 +1,0 @@
-'use client';
-
-import { FC } from 'react';
-import { useFormState } from 'react-dom';
-import {
-  TextInput,
-  HintText,
-  Label,
-  Textarea,
-  Button,
-  ChevronLeftIcon,
-} from 'nhsuk-react-components';
-import Link from 'next/link';
-import { processFormActions } from '@forms/CreateNhsAppTemplate/server-action';
-import { ZodErrorSummary } from '@molecules/ZodErrorSummary/ZodErrorSummary';
-import { NHSNotifyFormWrapper } from '@molecules/NHSNotifyFormWrapper/NHSNotifyFormWrapper';
-import { TemplateNameGuidance } from '@molecules/TemplateNameGuidance';
-import { Personalisation } from '@molecules/Personalisation/Personalisation';
-import { MessageFormatting } from '@molecules/MessageFormatting/MessageFormatting';
-import { NHSAppTemplate, PageComponentProps, Draft } from '@utils/types';
-import { TemplateType } from '@utils/enum';
-import { createNhsAppTemplatePageContent } from '@content/content';
-import { useTextInput } from '@hooks/use-text-input.hook';
-import { useJsEnabledStyle } from '@hooks/use-js-enabled-style.hook';
-
-export const CreateNhsAppTemplate: FC<
-  PageComponentProps<NHSAppTemplate | Draft<NHSAppTemplate>>
-> = ({ initialState }) => {
-  const {
-    pageHeading,
-    errorHeading,
-    buttonText,
-    characterCountText,
-    templateNameLabelText,
-    templateMessageLabelText,
-    templateNameHintText,
-  } = createNhsAppTemplatePageContent;
-  const [state, action] = useFormState(processFormActions, initialState);
-
-  const [nhsAppTemplateMessage, nhsAppMessageHandler] =
-    useTextInput<HTMLTextAreaElement>(state.message);
-
-  const [nhsAppTemplateName, nhsAppTemplateNameHandler] =
-    useTextInput<HTMLInputElement>(state.name);
-
-  const templateNameError =
-    state.validationError?.fieldErrors.nhsAppTemplateName?.join(', ');
-
-  const templateMessageError =
-    state.validationError?.fieldErrors.nhsAppTemplateMessage?.join(', ');
-
-  return (
-    <div className='nhsuk-grid-row'>
-      <div className='nhsuk-back-link nhsuk-u-margin-bottom-6 nhsuk-u-margin-left-3'>
-        <Link href='/choose-a-template-type' className='nhsuk-back-link__link'>
-          <ChevronLeftIcon />
-          Go back
-        </Link>
-      </div>
-      <div className='nhsuk-grid-column-two-thirds'>
-        <ZodErrorSummary errorHeading={errorHeading} state={state} />
-        <NHSNotifyFormWrapper action={action} formId='create-nhs-app-template'>
-          <h1 className='nhsuk-heading-xl' data-testid='page-heading'>
-            {pageHeading}
-          </h1>
-          <div className={templateNameError && 'nhsuk-form-group--error'}>
-            <Label htmlFor='nhsAppTemplateName'>{templateNameLabelText}</Label>
-            <HintText>{templateNameHintText}</HintText>
-            <TemplateNameGuidance template={TemplateType.NHS_APP} />
-            <TextInput
-              id='nhsAppTemplateName'
-              defaultValue={nhsAppTemplateName}
-              onChange={nhsAppTemplateNameHandler}
-              error={templateNameError}
-              errorProps={{ id: 'nhsAppTemplateName-error-message' }}
-            />
-          </div>
-          <Textarea
-            label={templateMessageLabelText}
-            id='nhsAppTemplateMessage'
-            maxLength={5000}
-            rows={10}
-            onChange={nhsAppMessageHandler}
-            defaultValue={nhsAppTemplateMessage}
-            error={templateMessageError}
-            errorProps={{ id: 'nhsAppTemplateMessage-error-message' }}
-          />
-          <p style={useJsEnabledStyle()} id='character-count'>
-            {nhsAppTemplateMessage.length}
-            {characterCountText}
-          </p>
-          <Button type='submit' id='create-nhs-app-template-submit-button'>
-            {buttonText}
-          </Button>
-        </NHSNotifyFormWrapper>
-      </div>
-      <div className='nhsuk-grid-column-one-third'>
-        <Personalisation />
-        <MessageFormatting template={TemplateType.NHS_APP} />
-      </div>
-    </div>
-  );
-};
